refactor: export DripsReceiverStruct and SplitsReceiverStruct from DripsHub
DripsReceiverStruct and SplitsReceiverStruct are exported from _all_ contracts, and it is confusing which type a user should import (although they are all the same - auto-generated).
By exporting like this, the user can import the necessary types directly from the library
instead of doing something like AddressDriver.DripsReceiverStruct, having to know and import the namespace.